backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
pdf2image
pytesseract
pillow
httpx
aiofiles
chromadb
sentence-transformers
numpy
scikit-learn
cryptography
tiktoken